Update why-storybook.mdx
Edited the Why Storybook page to make these changes:
- made the description of the initial problem more concise 
- unified the flow of the benefits section (before there were a lot of short sentences in some, with longer sentences in others) 
- improved the flow of the entire document 
- combined paragraphs 
Overall, I didn't change any of the content of the page, just tried to edit here and there to make the words flow a bit better. This is my first time contributing to an open-source project, so I hope this can help!