Removed 'other' option from example in viewport.mdx
While updating the configuration of viewports in project we decided to assign icons to our custom viewports in Storybook. That's when we spotted that assigning 'other' doesn't show the icon.
I've checked the names of the available icons in `Devices` object in `iconList` array, and 'others' doesn't seem to be there.