make StoryIndexGenerator do specifier lookup instead of watchStorySpecifier.
This is because now, the specifiers/normalized stories are being created in multiple places, so they are not referential equal. Therefore the specifier being passed in can't be used as a key to lookup the map in this.specifierToCache. But by just moving the importPathMatcher check into the StoryIndexGenerator, it will use its own specifiers, which _can_ be used as keys for looking up.

While this technically introduces a double pass on the specifiers, users don't have that many specifiers so it doesn't have any material impact on performance.